Switch to i128 instead of i64

+++ src/ast.rs
@@ -113,21 +113,21 @@ pub enum Token {
     #[token("<")]
     #[display(fmt = "<")]
     Lesser,
-    #[regex("[0-9][_0-9]*", |lex| i64::from_str_radix(&lex.slice().chars().filter(|&c| c != '_').collect::<String>(), 10).map_err(TokenError::from), priority = 2)]
+    #[regex("[0-9][_0-9]*", |lex| i128::from_str_radix(&lex.slice().chars().filter(|&c| c != '_').collect::<String>(), 10).map_err(TokenError::from), priority = 2)]
     #[regex("0?x[0-9a-fA-F][0-9a-fA-F_]*", |lex|
-        i64::from_str_radix(
+        i128::from_str_radix(
             &lex.slice().trim_start_matches("0x").trim_start_matches('x').chars().filter(|&c| c != '_').collect::<String>(),
             16,
         ).map_err(TokenError::from)
     )]
     #[regex("0?b[0-1][0-1_]*", |lex|
-        i64::from_str_radix(
+        i128::from_str_radix(
             &lex.slice().trim_start_matches("0b").trim_start_matches('b').chars().filter(|&c| c != '_').collect::<String>(),
             2,
         ).map_err(TokenError::from)
     )]
     #[display(fmt = "<number:{}>", _0)]
-    Number(i64),
+    Number(i128),
     #[regex("[0-9]+([eE][-+]?[0-9]+)?", |s| s.slice().parse(), priority = 1)]
     #[regex("\\.[0-9]+([eE][-+]?[0-9]+)?", |s| s.slice().parse())]
     #[regex("[0-9]+\\.[0-9]*([eE][-+]?[0-9]+)?", |s| s.slice().parse())]
@@ -162,7 +162,7 @@ impl std::fmt::Debug for Variable {
 }
 
 pub enum Literal {
-    Number(i64),
+    Number(i128),
     Float(f64),
     Atom(Arc<str>),
     Bool(bool),
@@ -337,7 +337,7 @@ pub enum UserParseError {
     #[error("Number '{num}' is out of range (expected '{ty}')")]
     OutOfRange {
         ty: &'static str,
-        num: i64,
+        num: i128,
         span: Range<usize>,
     },
 }
